Отрефакторены вспомогательные функции
