+++ src/components/event-edit.js
@@ -1,6 +1,6 @@
 import {generateOffers} from "../mock/offer.js";
 import {TRANSER_TYPES, ACTIVITY_TYPES, TYPE_PLACEHOLDER, DESTINATIONS} from "../const.js";
-import {formatTime} from "../utils.js";
+import {formatTime} from "../utils/datetime.js";
 import AbstractComponent from "./abstract-component.js";
 
 const createEventTypeItemMarkup = (type, isChecked) => {
