+++ src/components/event.js
@@ -1,5 +1,5 @@
 import {TYPE_PLACEHOLDER} from "../const.js";
-import {formatDate, formatTime, calculateDuration} from "../utils.js";
+import {formatDate, formatTime, calculateDuration} from "../utils/datetime.js";
 import AbstractComponent from "./abstract-component.js";
 
 const createScheduleMarkup = (start, stop) => {
