+++ src/mock/offer.js
@@ -1,4 +1,4 @@
-import {getRandomArrayItem, getRandomIntegerNumber} from "../utils.js";
+import {getRandomArrayItem, getRandomIntegerNumber} from "../utils/common.js";
 
 const OfferTypes = [`luggage`, `comfort`, `tickets`, `train`, `seats`, `lunch`, `breakfast`, `meal`];
 const OfferNames = [`Order Uber`, `Add luggage`, `Switch to comfort`,
