+++ src/utils/common.js
@@ -1,0 +1,8 @@
+export const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomIntegerNumber(0, array.length - 1);
+  return array[randomIndex];
+};
+
+export const getRandomIntegerNumber = (min, max) => {
+  return Math.floor(min + Math.random() * (max + 1 - min));
+};
